backend/requirements.txt
asgiref==3.4.1
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.6
Django==3.2.7
djangorestframework==3.12.4
idna==3.2
mysqlclient==2.0.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2021.3
requests==2.26.0
rsa==4.7.2
sqlparse==0.4.2
urllib3==1.26.7
uuid==1.30
